Move shouldBeAbleToDownloadDocuments method out to shared Page object
We will need to use this for the Assess journey so it is better
to share it
